+++ whiteboard-challenge-01/assets/array_reverse.JPG
